Fix minzoom, maxzoom and visibility handling
minzoom and maxzoom on layers now always refer to mapbox-gl zoom levels.
minzoom and mayzoom on sources refer to tilejson zoom levels and are
different from mapbox-gl zoom levels when the tile size is not 512.

This commit also changes visibility management from setting visible on
the layer when the resolution changes to setting minResolution and
maxResolution.