bugfix: Add request object to metadata if pending
fetchResource() caches pending requests (or rather the Promises of their
responses) to prevent repeated calls for the same resource. If a
response is not pending then the generated Request object may be
returned via the metadata parameter, but before this commit then if the
response was already pending the Request object was never returned in
this way. This led to problems in getTileJson() where it is expected
(required) that the request has been set.